+++ multiverse/themeoptions.php
@@ -19,9 +19,7 @@ class ThemeOptions {
     setThemeOptionDefault('social_content', ',,');
     setThemeOptionDefault('social_contacts', 0);
     setThemeOptionDefault('albums_per_page', 12);
-    setThemeOptionDefault('albums_per_row', 4);
     setThemeOptionDefault('images_per_page', 24);
-    setThemeOptionDefault('images_per_row', 2);
     setThemeOptionDefault('thumb_transition', 1);
     setThemeOptionDefault('thumb_transition_min', 24);
     setThemeOptionDefault('thumb_transition_max', 24);
@@ -59,24 +57,25 @@ class ThemeOptions {
         $unpublishedpages[get_language_string($page['title'])] = $page['titlelink'];
       }
     }
+    $me = "multiverse";
     return array(
-      gettext_th('Image width') => array(
+      gettext_th('Image width', $me) => array(
         'key' => 'image_size_x',
         'type' => OPTION_TYPE_CUSTOM,
         'order' => 1,
-        'desc' => gettext_th('Set the max width in pixels for the image displayed in the popup.')
+        'desc' => gettext_th('Set the max width in pixels for the image displayed in the popup.', $me)
       ),
-      gettext_th('Image height') => array(
+      gettext_th('Image height', $me) => array(
         'key' => 'image_size_y',
         'type' => OPTION_TYPE_CUSTOM,
         'order' => 2,
-        'desc' => gettext_th('Set the max height in pixels for the image displayed in the popup.') . '<p class="notebox">' . gettext_th("<strong>Note:</strong> images in the popup might be visually reduced in width and height to fit user screen, but they can be zoomed to the sizes set here by double clicking on them.") . '</p>'
+        'desc' => gettext_th('Set the max height in pixels for the image displayed in the popup.', $me) . '<p class="notebox">' . gettext_th("<strong>Note:</strong> images in the popup might be visually reduced in width and height to fit user screen, but they can be zoomed to the sizes set here by double clicking on them.", $me) . '</p>'
       ),
       gettext('Full image link') => array(
         'key' => 'full_image',
         'type' => OPTION_TYPE_CHECKBOX,
         'order' => 3,
-        'desc' => gettext_th('Check to show a link to the full size image on the image page. The behavior of the link depends on the options set in <code>Options->Image: Full image protection</code>.') . "</p>"
+        'desc' => gettext_th('Check to show a link to the full size image on the image page. The behavior of the link depends on the options set in <code>Options->Image: Full image protection</code>.', $me) . "</p>"
       ),
       gettext('Allow search') => array(
         'key' => 'search',
@@ -84,41 +83,41 @@ class ThemeOptions {
         'order' => 4,
         'desc' => gettext('Check to enable search form.')
       ),
-      gettext_th('Email subject') => array(
+      gettext_th('Email subject', $me) => array(
         'key' => 'email_subject',
         'type' => OPTION_TYPE_TEXTBOX,
         'order' => 5,
-        'desc' => gettext_th('Subject of the email sent through the contact form included in Multiverse footer. Zenphoto automatically adds the title of your gallery (between brackets)') . '<p class="notebox">' .  gettext_th('<strong>Note:</strong> leave this field empty to include a subject field in the contact form and let users fill it as they wish.') . "</p>"
+        'desc' => gettext_th('Subject of the email sent through the contact form included in Multiverse footer. Zenphoto automatically adds the title of your gallery (between brackets)', $me) . '<p class="notebox">' .  gettext_th('<strong>Note:</strong> leave this field empty to include a subject field in the contact form and let users fill it as they wish.', $me) . "</p>"
       ),
-      gettext_th('Copyright owner') => array(
+      gettext_th('Copyright owner', $me) => array(
         'key' => 'copyrigth_text',
         'type' => OPTION_TYPE_TEXTBOX,
         'order' => 6,
-        'desc' => gettext_th('Write the name of the copyright owner here. To this will be automatically added the year of the older image present in your gallery, followed by the current year, with a link to the full archive page.') . "<p class='notebox'>" . gettext_th("<strong>Note:</strong> leave this field empty if you don’t want any copyright text displayed.") . "</p>"
+        'desc' => gettext_th('Write the name of the copyright owner here. To this will be automatically added the year of the older image present in your gallery, followed by the current year, with a link to the full archive page.', $me) . "<p class='notebox'>" . gettext_th("<strong>Note:</strong> leave this field empty if you don’t want any copyright text displayed.", $me) . "</p>"
       ),
-      gettext_th('Copyright page') => array(
+      gettext_th('Copyright page', $me) => array(
         'key' => 'copyright_page',
         'type' => OPTION_TYPE_SELECTOR,
         'selections' => $allpages,
         'order' => 7,
         'null_selection' => gettext('none'),
-        'desc' => gettext_th("Choose the page that contains your site copyright terms, if any. A link to this page will be generated on the name of the copyright owner.") . "<p class='notebox'>" . gettext_th("<strong>Note:</strong> You need the Zenpage extension enabled to use this feature.") . "</p>"
+        'desc' => gettext_th("Choose the page that contains your site copyright terms, if any. A link to this page will be generated on the name of the copyright owner.", $me) . "<p class='notebox'>" . gettext_th("<strong>Note:</strong> You need the Zenpage extension enabled to use this feature.", $me) . "</p>"
       ),
-      gettext_th('Social profiles') => array(
+      gettext_th('Social profiles', $me) => array(
         'key' => 'social_contacts',
         'type' => OPTION_TYPE_CHECKBOX,
         'order' => 8,
-        'desc' => gettext_th("Check to display a list of icons with links to your social media profiles, defined by filling the fields below with the following data:")
+        'desc' => gettext_th("Check to display a list of icons with links to your social media profiles, defined by filling the fields below with the following data:", $me)
         . "<ul><li>"
-        . gettext_th("Your social profile URL")
+        . gettext_th("Your social profile URL", $me)
         . "</li><li>"
-        . gettext_th("Font Awesome class for the icon (i.e. <em>fa-github</em> for GitHub).")
+        . gettext_th("Font Awesome class for the icon (i.e. <em>fa-github</em> for GitHub).", $me)
         . '<a href="https://fontawesome.com/v4.7.0/icons/#brand" target="_blank" rel="noopener" title="FontAwesome 4.7"> '
-        . gettext_th("Here the full list") . "</a>"
+        . gettext_th("Here the full list", $me) . "</a>"
         . "</li><li>"
-        . gettext_th("Name of the social media")
+        . gettext_th("Name of the social media", $me)
         . "</li></ul>"
-        . gettext_th('Use the <strong>Add</strong> button to add a new social media or the <strong>Delete</strong> button to remove the last one.')
+        . gettext_th('Use the <strong>Add</strong> button to add a new social media or the <strong>Delete</strong> button to remove the last one.', $me)
       ),
       gettext('Homepage') => array(
         'key' => 'zenpage_homepage',
@@ -128,7 +127,7 @@ class ThemeOptions {
         'null_selection' => gettext('none'),
         'desc' => gettext("Choose here any <em>un-published Zenpage page</em> (listed by <em>titlelink</em>) to act as your site’s homepage instead the normal gallery index.")
         . "<p class='notebox'>"
-        . gettext_th("<strong>Note:</strong> You need the Zenpage extension enabled to use this feature.")
+        . gettext_th("<strong>Note:</strong> You need the Zenpage extension enabled to use this feature.", $me)
         . "</p>"
       ),
       gettext('News on index page') => array(
@@ -137,16 +136,16 @@ class ThemeOptions {
         'order' => 11,
         'desc' => gettext("Enable this if you want to show the news section’s first page on the <code>index.php</code> page.")
         . "<p class='notebox'>"
-        . gettext_th("<strong>Note:</strong> You need the Zenpage extension enabled to use this feature.")
+        . gettext_th("<strong>Note:</strong> You need the Zenpage extension enabled to use this feature.", $me)
         . "<br>"
-        . gettext_th("This overrides the <em>Homepage</em> option above.")
+        . gettext_th("This overrides the <em>Homepage</em> option above.", $me)
         . "</p>"
       ),
-      gettext_th('Enable Pagination') => array(
+      gettext_th('Enable Pagination', $me) => array(
         'key' => 'pagination',
         'type' => OPTION_TYPE_CHECKBOX,
         'order' => 12,
-        'desc' => gettext_th('Check to split albums into multiple pages, depending on the options <code>Albums</code> and <code>Images</code> found at the top of this page. The album design is cleaner with this option disabled (default), however if you have some albums with many images and you are experiencing a too long loading time, you may wish to enable this option.')
+        'desc' => gettext_th('Check to split albums into multiple pages, depending on the options <code>Albums</code> and <code>Images</code> found at the top of this page. The album design is cleaner with this option disabled (default), however if you have some albums with many images and you are experiencing a too long loading time, you may wish to enable this option.', $me)
       )
     );
   }